_resources/requirements.txt
pyyaml
git+https://github.com/sciunto-org/python-bibtexparser@main
